Move server signature code into supporting module
