+++ src/sasl.ts
@@ -29,3 +29,46 @@ export function hi(password: string, saltBytes: Buffer, iterations: number) {
     }
     return ui;
 }
+
+export function sign(data: string, password: string, clientNonce: string): [string, string] {
+    const m = Object.fromEntries(data.split(',').map(
+        (attr) => [attr[0], attr.substring(2)])
+    );
+
+    if (!(m.i && m.r && m.s)) throw new Error("SASL message parse error");
+
+    const nonce = m.r;
+
+    if (!nonce.startsWith(clientNonce))
+        throw new Error("SASL nonce mismatch");
+    if (nonce.length === clientNonce.length)
+        throw new Error("SASL nonce too short");
+
+    const iterations = parseInt(m.i, 10);
+    const salt = Buffer.from(m.s, 'base64');
+    const saltedPassword = hi(password, salt, iterations)
+
+    const clientKey = hmacSha256(saltedPassword, 'Client Key');
+    const storedKey = sha256(clientKey);
+
+    const clientFinalMessageWithoutProof = 'c=biws,r=' + nonce;
+    const clientFirstMessageBare = 'n=*,r=' + clientNonce;
+    const serverFirstMessage = data;
+
+    const authMessage = (
+        clientFirstMessageBare + ',' +
+        serverFirstMessage + ',' +
+        clientFinalMessageWithoutProof
+    );
+
+    const clientSignature = hmacSha256(storedKey, authMessage);
+    const clientProofBytes = xorBuffers(clientKey, clientSignature);
+    const clientProof = clientProofBytes.toString('base64');
+
+    const serverKey = hmacSha256(saltedPassword, 'Server Key');
+    const serverSignatureBytes = hmacSha256(serverKey, authMessage);
+
+    const response = clientFinalMessageWithoutProof + ',p=' + clientProof;
+    const serverSignature = serverSignatureBytes.toString('base64');
+    return [response, serverSignature];
+}
